// src/app/shared/models/request-models/user-request.model.ts
export class IUserRequestModel {
    id?: string;
    userName?: string;
    email?: string;
    password?: string;
    userID?: string;
    userType?: string;
    role?: string;
    roleId?:string;
    roleName?: string;
    mobileNo?: string;
    phoneNumber?: string;
    department?:string;
    departmentId?:any;
    claims?: [] = []
}